Added some customization to the cells.
This commit was to add some customization in terms of style to the UICollectionViewCells. The image that is shown when the user clicks on the cell is shown as a fully-covering image preview.

New class made for the custom cell and an imageView added to the cells contentView

Some modification to collectionView(_:cellForItemAtIndexPath) to make sure the cell has the right image passed to it from the cell to the imageView.