Completed regular material in Custom Cells Vid
This is the first part of the Custom Cells tutorial. In this code the cells were given a custom class and a background image that is the same as the indexPath.

The next upload will be completing the challange and adding a label to the cells with a gradient background so it can be read on any background.