Main part od "Multipil Cells" completed
In this commit there was a sub-class of UICollectionViewReusableView that set a title to the header.

The numbersOfSectionsInCollectionView and NumbeeOfItemsInSection were updated and an entirely new method was added named...

viewForSupplementaryElementOfKind...

This method setup a var that was set to dequeueReusableSupplimentaryViewOfKind, set the title to the section title and then returns the var from before.