Last commit for Section Headers:challange complete
In this commit there has been an image added to the Section header, the image names for the seasons are set to be the same as the title of the headers.

The image and text have been stacked to make their constraints work better.

Next commit will be the start of Inserting Cells...