Final commit for Deleting Cells and challange
In this commit we rolled-our-own delete batches of cells functionality. We did the following...

Added an image and set its autoconstraints which was a blank check circle for deleting. Connected it to PaperCell file.

Create editing proterty Bool.

override selected Bool.

Override the setEditing in MasterViewController (collectionview) and add a bunch of code to it.

Add one line to the cellForItemAtIndexPath method in MasterViewCntroller.

Change didSelectItemAtIndexPAth so that it doesn't transition into new screen.

Create didDeselectItemAtIndexPath to be the exact opposite of the above.

Set ToolbarHidden to true in the viewDidLoad.

Add bar button for the botton bar delete button and add spacers.

Connect the add button to an IBaction in MasterViewController